requirements.txt
attrs==21.2.0
jsonschema==3.2.0
pyrsistent==0.17.3
PyYAML==5.4.1
six==1.16.0
